- Typecast für Scheduler-Intervall

+++ src/main.cpp
@@ -71,10 +71,10 @@ Compass compFilterNorth(&status, 0, &status.magnetY, &status.magnetX, &status.an
             &status.filterCoefficientZ);
 
 PIDController pidControllerX(&status, PID_DEFAULT_PRIORITY,
-		SCHEDULER_INTERVALL_ms, &status.angleX, 0, &status.rcSignalNick, &status.pidXOut,
+		(float)SCHEDULER_INTERVALL_ms, &status.angleX, 0, &status.rcSignalNick, &status.pidXOut,
             0.15f, false);
 PIDController pidControllerY( &status, PID_DEFAULT_PRIORITY,
-		SCHEDULER_INTERVALL_ms, &status.angleY, 0, &status.rcSignalRoll, &status.pidYOut,
+		(float)SCHEDULER_INTERVALL_ms, &status.angleY, 0, &status.rcSignalRoll, &status.pidYOut,
 			0.15f, false);
 
 DiscoveryLEDs leds(&status, LEDs_DEFAULT_PRIORITY);
